Refactor clipboard history display to DataGridView
Replaced the ListView with a DataGridView for clipboard history items. Added new columns for "Type", "Content", and "Tag", and implemented a dropdown for tagging. Updated context menu functionality and added event handlers for tag editing and validation. Modified methods for saving text and images to ensure consistent functionality and improved user experience.

ToDo:
- Add posibility to create a new tab with a selection of tags
- Add default tab with favorite tags (default)