Add tag filtering to clipboard history
Introduces a `ComboBox` for tag filtering in the `ClipboardForm` class, allowing users to filter clipboard history items by tags. The `InitTagFilter` method initializes the ComboBox with tags from the database, including an "All" option. A new `GetByTag` method is added to the `Database` class to retrieve clipboard items based on the selected tag.

ToDo:
- Make Filter Tag dropdown better